updating to discord.js 1.12

+++ events/message/lvl.js
@@ -5,7 +5,7 @@ const prefix = fs.readFileSync("./config/prefix", "utf-8")
 
 exports.run = async (message, client) =>{
     if(message.content.includes(prefix)) return
-    if(message.member.roles.filter(role => role.name === "muted").size !==0) return
+    if(message.member.roles.cache.filter(role => role.name === "muted").size !==0) return
     if (fs.existsSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " lvl.json")){
         if (!fs.existsSync("./config/guild/" + message.guild.id + "/lvl/enabled")) return fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "off")
         if (!fs.existsSync("./config/guild/" + message.guild.id + "/lvl/enabled") || !(fs.readFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "utf-8") === "on")) return
@@ -32,7 +32,7 @@ exports.run = async (message, client) =>{
     }else{
         log.log("[JeuxGate : elvl] Creating lvl ", "messages", message.guild.id + "/lvl")
         if(fs.existsSync("./config/guild/" + message.guild.id + "/lvl/")){
-            fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "off")
+            if (!fs.existsSync("./config/guild/" + message.guild.id + "/lvl/enabled")) fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/enabled", "off")
             fs.writeFileSync("./config/guild/" + message.guild.id + "/lvl/" + message.author.id + " lvl.json", "{\"level\": 0,\"messagesend\": 0}")
         }else{
             fs.mkdirSync("./config/guild/" + message.guild.id + "/lvl/")
